feat: separate game logic from the ui
