+++ src/components/game/components/ui/pages/Game.tsx
@@ -5,7 +5,7 @@ import { GameHeader } from "../../../layouts/header/GameHeader";
 import { GameSidebarButtons } from "../../../layouts/sidebar/GameSidebar";
 import { AquariumTabs } from "../../../../aquarium/components/ui/tabs/AquariumTabs";
 import { TipsPopup } from "../tips/TipsPopup";
-import { FishDisplay } from "../display/FishDisplay";
+import { FishDisplay } from "../../../../aquarium/components/ui/display/FishDisplay";
 import { INITIAL_GAME_STATE, MOCK_FISH } from "../../../data/Game.data";
 import { useAquarium } from "../../../../aquarium/hooks/Aquarium.hook";
 import { useFishStats } from "../../../../aquarium/hooks/fish/FishStats.hook";
